requirements.txt
datasets
diffusers
transformers
timm
open-clip-torch
pyarrow
accelerate
einops
huggingface_hub[cli]
hf_transfer
tensorboard
matplotlib
protobuf
sentencepiece
tqdm
lovely_tensors
nvidia-nccl-cu12==2.20.5

